[Refactor] Move the lake PK index's row-oriented and parallel publish API onto the index itself
Both of the publish path's fan-out axes are supposed to be one class's business, but one of
them lived on LakePrimaryIndex and the other on LakePersistentIndex, which is much of why the
parallel code reads as tangled: the axis a caller drives (spread a rowset's chunks over the
token the publish supplies) and the internal one (split a batch of keys across key subsets and
sstable files on pk_index_execution_thread_pool) were in different files, so the rule that
binds them -- a task on the caller's token must not enter the internal axis, or the two pools
wait on each other -- was not stated anywhere.

Move parallel_get, batch_parallel_get_rss_rowids, parallel_upsert and upsert_owned onto
LakePersistentIndex, with owned_rowids_of, their one helper. Their bodies are unchanged; the
invariant above is now written down where both axes are visible.

They depend on the row-oriented wrappers -- the ones taking a rowset's primary-key Column and
the (rssid, rowid) coordinates its rows occupy, rather than an encoded Slice array -- so those
move too: get, three upserts, erase, bulk_erase, replace, try_replace. Each shrinks on the way
down. Every one opened by resolving the index it was about to delegate to and returning a
status if there was none; on LakePersistentIndex `this` is that index, and the encoded key size
is a member rather than a call back into the object being delegated to.

What is left on LakePrimaryIndex is that null check, as a forward per method. It is this class's
one remaining job beyond the load state and the publish lock, and it goes away with the class
when the two merge -- this is the step before that, so the merge can be a deletion rather than
a deletion plus a move.

One thing the forwards make visible: LakePrimaryIndex::erase and bulk_erase take a
TabletMetadataPtr that neither body ever read. The moved-down versions drop it; the forwards
still accept and ignore it, so the four call sites are untouched. Worth removing separately.
